# Improve performance of `case_when()` and `case_match()`