Closes https://github.com/tidyverse/dplyr/issues/6674

Three improvements to make these decently fast again:

- Minor: Calls to `current_env()` and `caller_env()` are now lazy (only evaluated when needed) (needs https://github.com/r-lib/rlang/issues/1558)

- Major: We no longer call `as_label()` to turn the expressions into labels which were being used as informative names that showed up in error messages. This stinks, but `as_label()` is just too slow to be used eagerly like this. We can revisit this again in the future if we add a lazy ALTREP character vector which could materialize the `as_label()` calls on demand (i.e. at error time). In the meantime, we now use positional names that also include LHS/RHS information, like `..2 (RHS)`, which ends up being decent.

- Major: I was previously wrapping the evaluation of the lhs/rhs of the formulas in `with_case_errors()` to give an informative indexed error about which formula failed to evaluate. This got slow as you increased the number of expressions in `case_when()` (2 * N expressions = number of `with_case_errors()` calls) because it uses `try_fetch()`, which is built on `tryCatch()`, which is in R not C (probably a good Luke project!). I've changed it up so 1 call to `with_case_errors()` now wraps the entire evaluation loop, and we track some index information to use in the error if one occurs.

Sadly the private option added in rlang (https://github.com/r-lib/rlang/commit/33db700d556b0b85a1fe78e14a53f95ac9248004) wasn't enough to make `as_label()` fast enough to be used eagerly like this. But the `infix_overflows()` call there is a big part of the performance degradation we see in my benchmarks below.

When I work on https://github.com/tidyverse/dplyr/issues/6678 that will end up also knocking off another 50us or so from switching away from `vec_assert()` in `vec_case_when()` and `if_else()`.

After that, any other improvements will come from moving `vec_case_when()` to C in vctrs.

Right now we are in the 2-4x slower range. After https://github.com/tidyverse/dplyr/issues/6678 I expect to be 2x slower most of the time. I think we can do a release with that. Then the C implementation in vctrs should fix the rest.

```r
library(dplyr)

# Size 1
x <- 1L

bench::mark(
  one = case_when(x == 1L ~ TRUE), 
  two = case_when(x == 1L ~ TRUE, x == 2L ~ NA),
  three = case_when(x == 1L ~ TRUE, x == 2L ~ NA, TRUE ~ FALSE),
  iterations = 50000,
  check = FALSE
)

# dplyr 1.0.10
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one          56.8µs   74.7µs    12802.     168KB     18.2
#> 2 two          75.5µs   95.4µs    10514.        0B     22.5
#> 3 three        97.4µs  120.2µs     8390.        0B     23.9

# This PR
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one           209µs    278µs     3388.   503.8KB     11.3
#> 2 two           242µs    326µs     2879.     1.3KB     12.2
#> 3 three         272µs    350µs     2787.     1.3KB     14.3

# dplyr 1.1.0 (smacked by the infix-overflow `as_label()` performance issue)
# (note that i had to go down to only 5,000 iterations here)
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one          1.67ms   2.12ms      477.    1.01MB     16.6
#> 2 two          3.14ms   3.86ms      261.    3.36KB     18.2
#> 3 three        3.27ms   4.32ms      228.    3.36KB     16.5
```

```r
library(dplyr)

# Size 1000
x <- 1:1000

bench::mark(
  one = case_when(x < 200 ~ TRUE), 
  two = case_when(x < 200 ~ TRUE, x < 500 ~ NA),
  three = case_when(x < 200 ~ TRUE, x < 500 ~ NA, TRUE ~ FALSE),
  iterations = 50000,
  check = FALSE
)

# dplyr 1.0.10
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one          73.2µs   92.2µs    10873.    60.1KB     17.0
#> 2 two         108.6µs  136.9µs     7383.   100.9KB     17.3
#> 3 three       136.1µs  171.2µs     5914.   130.6KB     18.3

# This PR
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one           222µs    271µs     3720.    44.1KB     14.4
#> 2 two           279µs    334µs     3011.    60.1KB     15.4
#> 3 three         318µs    382µs     2649.      70KB     16.5

# dplyr 1.1.0
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one          2.03ms   2.39ms      400.      45KB     15.0
#> 2 two          3.98ms   4.45ms      222.    62.2KB     16.6
#> 3 three         4.1ms   4.64ms      211.      72KB     16.6
```